+++ frontend/src/features/modals/AuthLoadingScreen.tsx
@@ -1,6 +1,6 @@
 import LoadingSpinner from "../../components/common/LoadingSpinner";
 import "../../../styles/modals/auth-loading-screen.scss";
-import ModalWrapper from "../../components/global/wrappers/ModalWrapper";
+import ModalWrapper from "../../components/wrappers/ModalWrapper";
 
 type AuthLoadingScreenProps = {
 	visible: boolean;
