+++ frontend/src/features/modals/MatchmakingShortcutScreen.tsx
@@ -1,7 +1,7 @@
 import { TimeControl } from "../../types/gameSetup";
 import { displayTimeControl } from "../../utils/timeUtils";
 import LoadingSpinner from "../../components/common/LoadingSpinner";
-import ModalWrapper from "../../components/global/wrappers/ModalWrapper";
+import ModalWrapper from "../../components/wrappers/ModalWrapper";
 
 import "../../../styles/modals/matchmaking-shortcut-screen.scss";
